cleaning up styles
