+++ src/app/page.tsx
@@ -1,9 +1,8 @@
 'use client'
 import Image from "next/image";
 import Navbar from "@/components/Navbar";
-// import AboutPage from "@/components/AboutPage";
-import React, { useState } from "react";
 import ContentWindow from "@/components/ContentWindow";
+import React, { useState } from "react";
 
 export default function Home() {
   const [content, setContent] = useState<string>('AboutPage');
@@ -15,14 +14,20 @@ export default function Home() {
           Carlton Bears Garden
         </p>
       </div>
-      <div className="flex height-full w-full fixed top-0">
-      <div className="items-end top-0 left-10 w-full bg-gradient-to-t from-white via-white dark:from-black dark:via-black lg:static lg:h-auto lg:w-auto lg:bg-none">
+      {/* <div className="flex height-full w-full fixed top-0"> */}
+      {/* <div className="fixed top-15 flex flex-col items-center justify-center w-1/2 h-25 bg-blue-500 "> */}
+      {/* <div className="flex flex-col items-center justify-center p-24 bg-blue-700 relative "> */}
+      <div>
+      {/* <div className="inline-block top-0 left-10 w-full bg-gradient-to-t from-white via-white dark:from-black dark:via-black lg:static lg:h-auto lg:w-auto lg:bg-none"> */}
           <Navbar content= {content} setContent={setContent} />
-        </div>
-        <div className="items-end top-50 left-50 h-1/4 w-1/2 bg-green-400">
-          <ContentWindow content={content} />
+        {/* </div> */}
+        {/* <div className="inline-block top-50 left-50 h-1/4 w-1/2 bg-green-400"> */}
+          {/* <div className="mt-10  "> */}
+          <div className="absolute bottom-0 left-0 right-0 mb-15">
+            <ContentWindow content={content} />
+          </div>  
           </div>
-      </div>
+      {/* </div> */}
       {/* <div className="relative flex place-items-center before:absolute before:h-[300px] before:w-full sm:before:w-[480px] before:-translate-x-1/2 before:rounded-full before:bg-gradient-radial before:from-white before:to-transparent before:blur-2xl before:content-[''] after:absolute after:-z-20 after:h-[180px] after:w-full sm:after:w-[240px] after:translate-x-1/3 after:bg-gradient-conic after:from-sky-200 after:via-blue-200 after:blur-2xl after:content-[''] before:dark:bg-gradient-to-br before:dark:from-transparent before:dark:to-blue-700 before:dark:opacity-10 after:dark:from-sky-900 after:dark:via-[#0141ff] after:dark:opacity-40 before:lg:h-[360px] z-[-1]"> */}
         {/* <Image
           className="relative dark:drop-shadow-[0_0_0.3rem_#ffffff70] dark:invert"
@@ -72,9 +77,9 @@ export default function Home() {
         <a
           href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
           className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
+          >
           target="_blank"
           rel="noopener noreferrer"
-        >
           <h2 className={`mb-3 text-2xl font-semibold`}>
             Templates{" "}
             <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
