--- src/components/TypeOfWindow.jsx
@@ -1,19 +1,0 @@
-// // ExampleComponent.js
-// import React from 'react';
-
-// const ExampleComponent = () => {
-//   // Check if window is defined (i.e., if code is running in the browser)
-//   const isServer = typeof window === 'undefined';
-
-//   return (
-//     <div>
-//       {isServer ? (
-//         <p>This code is running on the server.</p>
-//       ) : (
-//         <p>This code is running on the client (browser).</p>
-//       )}
-//     </div>
-//   );
-// };
-
-// export default ExampleComponent;
