--- src/content/about.mdx
@@ -1,1 +1,0 @@
-'test'
